requirements.txt
fastapi
uvicorn
chromadb
ollama
pydantic
pdfminer.six
python-multipart